+++ Cargo.toml
@@ -9,7 +9,7 @@ edition = "2018"
 [dependencies]
 bitcoincore-rpc = "0.13"
 #bitcoincore-rpc = {git="https://github.com/rust-bitcoin/rust-bitcoincore-rpc"}
-bitcoin-wallet = "1.1.0"
+bip39 = {version = "1.0.1", features = ["rand"] }
 bitcoin = {version = "0.26", features = ["rand"] }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
