+++ src/error.rs
@@ -1,5 +1,7 @@
 use std::{error, io};
 
+use crate::{market::directory::DirectoryServerError, wallet::WalletError};
+
 // error enum for the whole project
 // try to make functions return this
 #[derive(Debug)]
@@ -9,6 +11,8 @@ pub enum TeleportError {
     Protocol(&'static str),
     Rpc(bitcoincore_rpc::Error),
     Socks(tokio_socks::Error),
+    Wallet(WalletError),
+    Market(DirectoryServerError),
 }
 
 impl From<Box<dyn error::Error + Send>> for TeleportError {
@@ -32,5 +36,17 @@ impl From<bitcoincore_rpc::Error> for TeleportError {
 impl From<tokio_socks::Error> for TeleportError {
     fn from(e: tokio_socks::Error) -> TeleportError {
         TeleportError::Socks(e)
+    }
+}
+
+impl From<WalletError> for TeleportError {
+    fn from(value: WalletError) -> Self {
+        Self::Wallet(value)
+    }
+}
+
+impl From<DirectoryServerError> for TeleportError {
+    fn from(value: DirectoryServerError) -> Self {
+        Self::Market(value)
     }
 }
