+++ src/protocol/mod.rs
@@ -1,0 +1,4 @@
+pub mod contract;
+pub mod messages;
+
+pub use contract::Hash160;
